pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ruleshield"
version = "0.2.0"
description = "Open-source LLM cost optimizer -- 4-layer intelligence proxy that reduces API costs by 50-80%"
authors = [{name = "RuleShield Contributors"}]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.10"
keywords = ["llm", "cost-optimization", "proxy", "cache", "openai", "anthropic", "agent", "hermes"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "httpx>=0.25.0",
    "openai>=1.3.0",
    "rich>=13.0.0",
    "sentence-transformers>=2.2.0",
    "click>=8.1.0",
    "pyyaml>=6.0",
    "aiosqlite>=0.19.0",
    "numpy>=1.24.0",
]

[project.urls]
Homepage = "https://github.com/banse/ruleshield-hermes"
Documentation = "https://github.com/banse/ruleshield-hermes#readme"
Repository = "https://github.com/banse/ruleshield-hermes"
"Issue Tracker" = "https://github.com/banse/ruleshield-hermes/issues"

[project.optional-dependencies]
dev = ["pytest", "pytest-asyncio", "ruff"]

[project.scripts]
ruleshield = "ruleshield.cli:main"

[tool.setuptools.packages.find]
include = ["ruleshield*"]
